requirements.txt
asgiref==3.3.4
attrs==21.4.0
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Deprecated==1.2.12
Django==3.2.4
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
docopt==0.6.2
google==3.0.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud==0.34.0
google-cloud-speech==2.11.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
Js2Py==0.71
jwcrypto==1.0
libcst==0.3.23
lxml==4.8.0
mailjet-rest==1.3.4
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
parsel==1.6.0
pipwin==0.5.1
priority==1.3.0
Protego==0.2.1
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyjsparser==2.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.6
PyPrind==2.11.3
pySmartDL==1.3.4
python-scrapyd-api==2.1.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
queuelib==1.6.2
requests==2.26.0
rsa==4.8
Scrapy==2.5.1
scrapyd==1.3.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
Twisted==22.1.0
twisted-iocpsupport==1.0.2
typing-inspect==0.7.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
w3lib==1.22.0
whitenoise==5.3.0
wrapt==1.12.1
zope.interface==5.4.0
